[core] Widen health code to 64-bit
Changing the health code proto type from uint32 to uint64 is wire-safe.

Details here: https://protobuf.dev/programming-guides/proto3/

+++ trellis/core/health.hpp
@@ -42,7 +42,7 @@ class Health {
   using HealthPublisher = trellis::core::Publisher<trellis::core::HealthHistory>;
   using PublisherCreateFunction = std::function<HealthPublisher(const std::string&)>;
   using HealthHistory = std::deque<trellis::core::HealthStatus>;
-  using Code = uint32_t;
+  using Code = uint64_t;
 
   /**
    * Health construct the health object
